Move InserterPoint to ControlPoints
